feat: Add mob anim. & touchFunc
Add enemies animation. Write definition for
touch function in landscape class

+++ Classes/TDGame/include/Cell.h
@@ -37,11 +37,11 @@ namespace TowerDefence{
 		/**
 		\return Cell height
 		*/
-		size_t getHeight()const;
+		float getHeight()const;
 		/**
 		\return Cell width
 		*/
-		size_t getWidth()const;
+		float getWidth()const;
 		/**
 		\return Top level point of cell
 		*/
